Move files back out of init.
